requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
fake-useragent==1.5.1
fonttools==4.53.0
Pillow==10.3.0
ddddocr==1.5.5
parsel==1.9.1